% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-checks.R
\name{accept_content}
\alias{accept_content}
\title{Run clean-up tasks and return built dataset}
\usage{
accept_content(results, df, blueprint, meta)
}
\arguments{
\item{results}{A reference to the checks results. Currently used to
ensure that this step runs after the checks step.}

\item{df}{The built dataset}

\item{blueprint}{The blueprint associated with the built dataset}

\item{meta}{The metadata associated with the built dataset}
}
\description{
After checks pass, this step runs in the blueprint sequence.
If any cleanup features are enabled, they will run on the
dataset prior to setting the final blueprint target.
}
