% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{attach_blueprints}
\alias{attach_blueprints}
\alias{attach_blueprint}
\title{Attach blueprints to a drake plan}
\usage{
attach_blueprints(plan, ...)

attach_blueprint(plan, blueprint)
}
\arguments{
\item{plan}{A drake plan}

\item{...}{Multiple blueprints}

\item{blueprint}{A blueprint object}
}
\description{
Blueprints outline a sequence of checks and cleanup steps
that come after a dataset is created. In order for these
steps to be executed, the blueprint must be attached to
a drake plan so that drake can run these steps properly.
}
