% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint}
\alias{blueprint}
\title{Create a blueprint}
\usage{
blueprint(
  name,
  command,
  description = NULL,
  metadata = NULL,
  export_metadata = TRUE,
  metadata_file_type = c("csv"),
  metadata_directory = here::here("blueprints"),
  metadata_file_path = NULL,
  ...,
  class = character()
)
}
\arguments{
\item{name}{The name of the blueprint}

\item{command}{The code to build the target dataset}

\item{description}{An optional description of the dataset to be used for
codebook generation}

\item{metadata}{The associated variable metadata for this dataset}

\item{export_metadata}{Indicator to render new metadata file for new dataset}

\item{class}{}
}
\value{

}
\description{
Create a blueprint
}
