% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint_target_name}
\alias{blueprint_target_name}
\alias{blueprint_target_name.default}
\alias{blueprint_target_name.character}
\alias{blueprint_target_name.blueprint}
\alias{blueprint_checks_name}
\alias{blueprint_checks_name.default}
\alias{blueprint_checks_name.character}
\alias{blueprint_checks_name.blueprint}
\alias{blueprint_final_name}
\alias{blueprint_final_name.default}
\alias{blueprint_final_name.character}
\alias{blueprint_final_name.blueprint}
\alias{blueprint_reference_name}
\alias{blueprint_reference_name.default}
\alias{blueprint_reference_name.character}
\alias{blueprint_reference_name.blueprint}
\alias{blueprint_codebook_name}
\alias{blueprint_codebook_name.default}
\alias{blueprint_codebook_name.character}
\alias{blueprint_codebook_name.blueprint}
\title{Blueprint drake target names}
\usage{
blueprint_target_name(x, ...)

\method{blueprint_target_name}{default}(x, ...)

\method{blueprint_target_name}{character}(x, ...)

\method{blueprint_target_name}{blueprint}(x, ...)

blueprint_checks_name(x, ...)

\method{blueprint_checks_name}{default}(x, ...)

\method{blueprint_checks_name}{character}(x, ...)

\method{blueprint_checks_name}{blueprint}(x, ...)

blueprint_final_name(x, ...)

\method{blueprint_final_name}{default}(x, ...)

\method{blueprint_final_name}{character}(x, ...)

\method{blueprint_final_name}{blueprint}(x, ...)

blueprint_reference_name(x, ...)

\method{blueprint_reference_name}{default}(x, ...)

\method{blueprint_reference_name}{character}(x, ...)

\method{blueprint_reference_name}{blueprint}(x, ...)

blueprint_codebook_name(x, ...)

\method{blueprint_codebook_name}{default}(x, ...)

\method{blueprint_codebook_name}{character}(x, ...)

\method{blueprint_codebook_name}{blueprint}(x, ...)
}
\arguments{
\item{x}{A blueprint or string relating to a blueprint's name}

\item{...}{Unused for now}
}
\description{
Blueprint drake target names
}
