% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprintr-package.R
\docType{package}
\name{blueprintr-package}
\alias{blueprintr}
\alias{blueprintr-package}
\title{blueprintr: Automagically Document and Test Datasets Using Targets Or Drake}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Documents and tests datasets in a reproducible manner so that data lineage is easier to comprehend for small to medium tabular data. Originally designed to aid data cleaning tasks for humanitarian research groups, specifically large-scale longitudinal studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nyuglobalties/blueprintr}
  \item Report bugs at \url{https://github.com/nyuglobalties/blueprintr/issues}
}

}
\author{
\strong{Maintainer}: Patrick Anker \email{psanker@nyu.edu} (\href{https://orcid.org/0000-0003-2302-0445}{ORCID})

Other contributors:
\itemize{
  \item Hillary Gao [contributor]
  \item Global TIES for Children [copyright holder]
}

}
\keyword{internal}
