% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-base.R
\name{bp_add_bpstep}
\alias{bp_add_bpstep}
\title{Add custom bpstep to blueprint schema}
\usage{
bp_add_bpstep(bp, step)
}
\arguments{
\item{bp}{A blueprint}

\item{step}{A bpstep object}
}
\description{
\code{blueprint()} objects store custom \link{bpstep} objects
in the "extra_steps" element. This function adds a new
step to that element.
}
\examples{
if (FALSE) {
  # Based on the codebook export step
  step <- bpstep(
    step = "export_codebook",
    bp = bp,
    payload = bpstep_payload(
      target_name = blueprint_codebook_name(bp),
      target_command = codebook_export_call(bp),
      format = "file",
      ...
    )
  )

  bp_add_bpstep(
    bp,
    step
  )
}
}
