% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint-extensions.R
\name{bp_export_codebook}
\alias{bp_export_codebook}
\title{Instruct blueprint to export codebooks}
\usage{
bp_export_codebook(
  blueprint,
  summaries = FALSE,
  file = NULL,
  template = NULL,
  title = NULL
)
}
\arguments{
\item{blueprint}{A blueprint}

\item{summaries}{Whether or not variable summaries should be included in codebook}

\item{file}{Path to where the codebook should be saved}

\item{template}{A path to an RMarkdown template}

\item{title}{Optional title of codebook}
}
\value{
An amended blueprint with the codebook export instructions
}
\description{
Instruct blueprint to export codebooks
}
\examples{
\dontrun{
test_bp <- blueprint(
  "mtcars_dat",
  description = "The mtcars dataset",
  command = mtcars
)

new_bp <- test_bp \%>\% bp_export_codebook()
}
}
