% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-export_kfa_report.R
\name{bp_export_kfa_report}
\alias{bp_export_kfa_report}
\title{Instruct blueprint to generate kfa report}
\usage{
bp_export_kfa_report(
  bp,
  scale,
  path = NULL,
  path_pattern = NULL,
  format = NULL,
  title = NULL,
  kfa_args = list(),
  ...
)
}
\arguments{
\item{bp}{A blueprint}

\item{scale}{Which scale(s) to analyze}

\item{path}{Path(s) to where the report(s) should be saved}

\item{path_pattern}{Override the default location to save files
(always rooted to the project root with here::here())}

\item{format}{The output format of the report(s)}

\item{title}{Optional title of report}

\item{kfa_args}{Arguments forwarded to \code{kfa::kfa()} for this batch
of scales}

\item{...}{Arguments forwarded to the executing engine
e.g. targets::tar_target_raw() or drake::target()}
}
\value{
An amended blueprint with the kfa report export instructions
}
\description{
Instruct blueprint to generate kfa report
}
\examples{
\dontrun{
test_bp <- blueprint(
  "mtcars_dat",
  description = "The mtcars dataset",
  command = mtcars
)

new_bp <- test_bp \%>\% bp_export_codebook()
}
}
