% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-base.R
\name{bp_extend}
\alias{bp_extend}
\title{Add custom elements to a blueprint}
\usage{
bp_extend(bp, ...)
}
\arguments{
\item{bp}{A blueprint}

\item{...}{Keyword arguments forwarded to blueprint()}
}
\description{
\code{blueprint()} objects are essentially just \code{list()} objects
that contain a bunch of metadata on the data asset construction.
Use \code{bp_extend()} to set or add new elements.
}
\examples{
if (FALSE) {
  bp <- blueprint("some_blueprint", ...)
  adjusted_bp <- bp_extend(bp, new_option = TRUE)
  bp_with_annotation_set <- bp_extend(bp, annotate = TRUE)
}
}
