% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-base.R
\name{bp_include_panelcleaner_meta}
\alias{bp_include_panelcleaner_meta}
\title{Include panelcleaner mapping on metadata creation}
\usage{
bp_include_panelcleaner_meta(blueprint)
}
\arguments{
\item{blueprint}{A blueprint that may create a \code{mapped_df} data.frame}
}
\value{
An amended blueprint with \code{mapped_df} metadata extraction set
for metadata creation
}
\description{
\href{https://nyuglobalties.github.io/panelcleaner/]}{panelcleaner} defines
a mapping structure used for data import of panel, or more generally
longitudinal, surveys / data which can be used as a source for some
kinds of metadata (currently, only categorical coding information).
If the blueprint constructs a \code{mapped_df} object, then this extension
will signal to blueprintr to extract the mapping information and
include it.
}
