% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint-extensions.R
\name{bp_label_variables}
\alias{bp_label_variables}
\title{Convert variables to labelled variables in cleanup stage}
\usage{
bp_label_variables(blueprint)
}
\arguments{
\item{blueprint}{A blueprint}
}
\value{
An amended blueprint with variable labelling in the cleanup phase set
}
\description{
The \href{https://haven.tidyverse.org/}{haven} package has a handy tool called
"labeled vectors", which are like factors that can be interpreted in other
statistical software like STATA and SPSS. See \link[haven:labelled]{haven::labelled()}
for more information on the type. Running this on a blueprint will instruct
the blueprint to convert all variables with non-NA \code{title}, \code{description}, or
\code{coding} fields to labeled vectors.
}
