% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpstep.R
\name{bpstep}
\alias{bpstep}
\title{Define a step of blueprint assembly}
\usage{
bpstep(step, bp, payload, ..., .class = NULL)
}
\arguments{
\item{step}{The name of the step}

\item{bp}{A 'blueprint' object to create the assembled step}

\item{payload}{The assembled object, either a 'drake_plan' or 'tar_target'}

\item{...}{Any extra information. Currently unused.}

\item{.class}{A character string to subclass \code{bpstep()}}
}
\value{
A 'bpstep' object
}
\description{
Each step in the blueprint assembly process is contained in a wrapper
'bpstep' object. \code{bpstep()} is the abstract class. \code{drake_bpstep()} and
\code{targets_bpstep()} are the implementations used within the \verb{bpstep_*()}
methods.
}
\keyword{internal}
