% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpstep.R
\name{bpstep_payload}
\alias{bpstep_payload}
\title{Create a step payload}
\usage{
bpstep_payload(target_name, target_command, ...)
}
\arguments{
\item{target_name}{The target's name}

\item{target_command}{The target's command}

\item{...}{Arguments to be passed to the executing engine (e.g.
arguments sent to targets::tar_target())}
}
\value{
A bpstep payload object
}
\description{
The bpstep payload is the object that contains the target name
and command, along with any other metadata to be passed to the
execution engine.
}
\examples{
if (FALSE) {
  bpstep(
    step = "some_step",
    bp = some_bp_object,
    payload = bpstep_payload(
      "payload_name",
      payload_command()
    )
  )
}
}
