% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-base.R
\name{check_list}
\alias{check_list}
\title{Create a quoted list of check calls}
\usage{
check_list(...)
}
\arguments{
\item{...}{A collection of calls to be used for checks}
}
\description{
Create a quoted list of check calls
}
