% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\alias{all_variables_present}
\alias{all_types_match}
\title{Evaluate checks on the blueprint build output}
\usage{
all_variables_present(df, meta, blueprint)

all_types_match(df, meta)
}
\arguments{
\item{df}{The built dataset}

\item{meta}{The dataset's metadata}

\item{blueprint}{The dataset's blueprint}
}
\description{
After building a dataset, it's beneficial (if not
a requirement) to run tests on that dataset to ensure
that it behaves as expected. \code{blueprintr} gives authors
a framework to run these tests automatically, both for
individual variables and general dataset checks.
\code{blueprintr} provides three functions as models for developing
these kinds of functions: one to check that all expected variables
are present, one to check the variable types, and a generic
function that checks if variable values are contained within
a known set.
}
