% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-create.R
\name{create_metadata_file}
\alias{create_metadata_file}
\title{Create a metadata file from a dataset}
\usage{
create_metadata_file(df, blueprint, ...)
}
\arguments{
\item{df}{A dataframe that the metadata table describes}

\item{blueprint}{The original blueprint for the dataframe}

\item{...}{A variable list of metadata tables on which this
metadata table depends}
}
\description{
One of the targets in the blueprint workflow target chain.
If a metadata file does not exist, then this function will be
added to the workflow.
}
