% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-base.R
\name{eval_checks}
\alias{eval_checks}
\title{Evaluate all checks on a blueprint}
\usage{
eval_checks(..., .env = parent.frame())
}
\arguments{
\item{...}{All quoted check calls}

\item{.env}{The environment in which the calls are evaluated}
}
\description{
Runs all checks -- dataset and variable -- on a blueprint
to determine if a built dataset passes all restrictions.
}
\section{Check functions}{
Check functions are simple functions that take in either
a data.frame or variable at the minimum, plus some extra
arguments if need, and returns a logical value: \code{TRUE} or \code{FALSE.}
In blueprintr, the entire check passes or fails unlike other
testing frameworks like {pointblank}. If you'd like to embed
extra context for your test result, modify the "check.errors"
attribute of the returned logical value with a character vector
which will be rendered into a bulleted list. Note: if you embed
reasons for a \code{TRUE}, the check will produce a warning in the drake
or targets pipeline.
}

