% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint-provenance.R
\name{get_table_linage_igraph}
\alias{get_table_linage_igraph}
\title{Get an igraph of the table lineage}
\usage{
get_table_linage_igraph(blueprints)
}
\arguments{
\item{blueprints}{a list() of blueprint objects}
}
\value{
An igraph object of the table lineage structure
}
\description{
Get an igraph of the table lineage
}
