% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage-variables.R
\name{get_variable_linage_igraph}
\alias{get_variable_linage_igraph}
\title{Get an igraph of the variable lineage}
\usage{
get_variable_linage_igraph(blueprints)
}
\arguments{
\item{blueprints}{a list() of blueprint objects}
}
\value{
An igraph object of the variable lineage structure
}
\description{
Get an igraph of the variable lineage
}
