% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-blueprints.R
\name{load_blueprint}
\alias{load_blueprint}
\alias{load_blueprints}
\title{Load a blueprint from a script file}
\usage{
load_blueprint(plan, file)

load_blueprints(plan, directory = here::here("blueprints"), recurse = FALSE)
}
\arguments{
\item{plan}{A drake plan}

\item{file}{A path to a script file}

\item{directory}{A path to a directory with script files that are blueprints.
Defaults to the "blueprints" directory at the root of the
current R project.}

\item{recurse}{Recursively loads blueprints from a directory if \code{TRUE}}
}
\value{
A drake_plan with attached blueprints
}
\description{
Load a blueprint from a script file
}
\section{Empty blueprint folder}{
By default, blueprintr ignore empty blueprint folders. However, it may be beneficial
to warn users if folder is empty, particularly during project setup. This helps
identify any potential misconfiguration of drake plan attachment. To enable these warnings,
set \code{option(blueprintr.warn_empty_blueprints_dirs = TRUE)}.
}

