% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Convert an input dataframe into a metadata object}
\usage{
metadata(df)
}
\arguments{
\item{df}{A dataframe that will be converted into a
metadata object, once content checks pass.}
}
\description{
Convert an input dataframe into a metadata object
}
