% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_from_blueprint}
\alias{plan_from_blueprint}
\title{Create a drake plan from a blueprint}
\usage{
plan_from_blueprint(blueprint)
}
\arguments{
\item{blueprint}{A blueprint}
}
\value{
A drake plan with all of the necessary blueprint steps
}
\description{
Creates a new drake plan from a blueprint
}
