% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read the metadata file}
\usage{
read_metadata(metadata_file)
}
\arguments{
\item{metadata_file}{Path of metadata file. Currently only supports CSVs}
}
\value{
A \code{metadata()} object
}
\description{
Reads the metadata from, found at a given path, and returns a \code{metadata()}
object. This is mostly an internal function for the blueprint steps.
}
\keyword{internal}
