% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-codebook.R
\name{render_codebook}
\alias{render_codebook}
\title{Render codebooks for datasets}
\usage{
render_codebook(
  blueprint,
  meta,
  file,
  title = glue("{ui_value(blueprint$name)} Codebook"),
  dataset = NULL,
  template = bp_path("codebook_templates/default_codebook.Rmd"),
  ...
)
}
\arguments{
\item{blueprint}{A dataset blueprint}

\item{meta}{A \code{blueprint_metadata} object related to the blueprint}

\item{file}{Path to where the codebook should be saved}

\item{title}{Title of the codebook}

\item{dataset}{If included, a \code{data.frame} to be used as a source for
summaries}

\item{template}{Path to the knitr template}

\item{...}{Extra parameters passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\description{
Render codebooks for datasets
}
