% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-export_kfa_report.R
\name{render_kfa_report}
\alias{render_kfa_report}
\title{Render k-fold factor analysis on scale using kfa}
\usage{
render_kfa_report(dat, meta, scale, path = NULL, format = NULL, title = NULL)
}
\arguments{
\item{dat}{Source data}

\item{meta}{blueprintr data dictionary}

\item{scale}{Scale identifier to be located in the 'scale' field}

\item{path}{Where to output the report; defaults to the "reports"
subfolder of the current working \emph{project} folder.}

\item{format}{The output format; defaults to 'html_document'}

\item{title}{Optional title of the report}
}
\value{
Path to where the generated report is saved
}
\description{
Generates a k-fold factor analys report using the 'scale'
field in the blueprintr data dictionaries. While not recommended,
this function does allow for multiple loaded variables, delimited by
commas. For example, 'var1' could have 'scale' be "SCALE1,SCALE2".
}
