% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotating.R
\name{super_annotations}
\alias{super_annotations}
\alias{improved_annotation_option}
\alias{using_improved_annotations}
\title{"Super Annotations"}
\usage{
improved_annotation_option()

using_improved_annotations()
}
\description{
As of blueprintr 0.2.1, there is now the option for metadata files to
\strong{always} overwrite annotations at runtime. Previously, this would be
a conflict with \link{mutate_annotation} and \link{mutate_annotation_across}
since the annotation phase happens during the blueprint cleanup phase, whereas these
annotation manipulation tools occur at the blueprint initial phase. To resolve
this, 0.2.1 introduces "super annotations", which are just annotations prefixed
with "super.". However, the super annotations will \emph{overwrite} the normal annotations
during cleanup. This gives the annotation manipulation tools a means of not losing their
work if \code{annotate_overwrite} is effectively enabled. To enable this functionality,
set \code{options(blueprintr.use_improved_annotations = TRUE)}. This also has the side effect
of \strong{always} treating \code{annotate = TRUE} and \code{annotate_overwrite = TRUE}.
}
\section{Functions}{
\itemize{
\item \code{improved_annotation_option()}: Returns the option string for improved annotations

\item \code{using_improved_annotations()}: Checks if improved annotations are enabled

}}
