% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage-variables.R
\name{variable_lineage}
\alias{variable_lineage}
\alias{load_variable_lineage}
\alias{vis_variable_lineage}
\title{Variable lineage}
\usage{
load_variable_lineage(
  directory = here::here("blueprints"),
  recurse = FALSE,
  script = here::here("_targets.R")
)

vis_variable_lineage(..., g = NULL, cluster_by_dataset = TRUE)
}
\arguments{
\item{directory}{A folder containing blueprint scripts}

\item{recurse}{Should this function recursively load blueprints?}

\item{script}{Where the targets/drake project script file is located. Defaults
to using targets.}

\item{...}{Arguments passed to \link{load_variable_lineage}}

\item{g}{An igraph object. This defaults to a graph loaded with
\link{load_variable_lineage}. However, use this if you
want to inspect subgraphs of the variable lineage.}

\item{cluster_by_dataset}{If \code{TRUE}, variable nodes will be clustered into their
respective dataset}
}
\description{
This is an experimental feature that traces variable lineage through
an injection of a ".uuid" attribute for each variable. Previous attempts
at variable lineage were conducted using variable names and heuristics
of known functions. This approach yields a more consistent lineage.
}
\details{
To enable the variable feature, set \code{options(blueprintr.use_variable_uuids = TRUE)}.
}
\section{Functions}{
\itemize{
\item \code{load_variable_lineage()}: Reads blueprintrs from folder to get variable lineage.
Returns an igraph of the variable lineage.

\item \code{vis_variable_lineage()}: Visualizes variable lineage with visNetwork.
Returns an interactive graph.

}}
