% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotating.R
\name{annotations}
\alias{annotations}
\alias{annotation_names}
\alias{annotation}
\alias{has_annotation}
\alias{add_annotation}
\title{Access the blueprintr metadata at runtime}
\usage{
annotations(x)

annotation_names(x)

annotation(x, field)

has_annotation(x, field)

add_annotation(x, field, value, overwrite = FALSE)
}
\arguments{
\item{x}{An object, most likely a variable in a \code{data.frame}}

\item{field}{The name of a metadata field}

\item{value}{A value to assign to an annotation field}

\item{overwrite}{If \code{TRUE}, allows overwriting of existing
annotation values}
}
\description{
Access the blueprintr metadata at runtime
}
\section{Functions}{
\itemize{
\item \code{annotations()}: Gets a list of all decorations assigned to an object

\item \code{annotation_names()}: Get the names of all of the
decorations assigned to an object

\item \code{annotation()}: Gets a decoration for an object

\item \code{has_annotation()}: Checks to see if a decoration exists for an object

\item \code{add_annotation()}: Adds a decoration to an object,
with the option of overwriting an existing value

}}
