% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint}
\alias{blueprint}
\title{Create a blueprint}
\usage{
blueprint(
  name,
  command,
  description = NULL,
  metadata = NULL,
  export_metadata = TRUE,
  metadata_file_type = c("csv"),
  metadata_directory = here::here("blueprints"),
  metadata_file_path = NULL,
  ...,
  class = character()
)
}
\arguments{
\item{name}{The name of the blueprint}

\item{command}{The code to build the target dataset}

\item{description}{An optional description of the dataset to be used for
codebook generation}

\item{metadata}{The associated variable metadata for this dataset}

\item{export_metadata}{Indicator to render new metadata file for new dataset}

\item{metadata_file_type}{The kind of metadata file. Currently only CSV.}

\item{metadata_directory}{Where the metadata file will be stored.}

\item{metadata_file_path}{Overrides the metadata file path generated by
\code{metadata_directory}, \code{name}, and \code{metadata_file_type}
if not NULL.}

\item{...}{Any other parameters and settings for the blueprint}

\item{class}{A subclass of blueprint capability, for future work}
}
\value{
A blueprint object
}
\description{
Create a blueprint
}
