% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-base.R
\name{bp_add_bpstep}
\alias{bp_add_bpstep}
\title{Add custom bpstep to blueprint schema}
\usage{
bp_add_bpstep(bp, step)
}
\arguments{
\item{bp}{A blueprint}

\item{step}{A bpstep object}
}
\description{
Add custom bpstep to blueprint schema
}
