% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-base.R
\name{bp_extend}
\alias{bp_extend}
\title{Add custom elements to a blueprint}
\usage{
bp_extend(bp, ...)
}
\arguments{
\item{bp}{A blueprint}

\item{...}{Keyword arguments forwarded to blueprint()}
}
\description{
Add custom elements to a blueprint
}
