% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpstep.R
\name{bpstep}
\alias{bpstep}
\title{Define a step of blueprint assembly}
\usage{
bpstep(step, bp, payload, ...)
}
\arguments{
\item{step}{The name of the step}

\item{bp}{A 'blueprint' object to create the assembled step}

\item{payload}{A 'bpstep_payload' object that outlines the code
to be assembled depending on the workflow executor}

\item{...}{Extensions to the bpstep, like "allow_duplicates"}
}
\value{
A 'bpstep' object
}
\description{
Each step in the blueprint assembly process is contained in a wrapper
'bpstep' object.
}
\keyword{internal}
