% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{create_metadata_file}
\alias{create_metadata_file}
\title{Create a metadata file from a dataset}
\usage{
create_metadata_file(df, blueprint, ...)
}
\arguments{
\item{df}{A dataframe that the metadata table describes}

\item{blueprint}{The original blueprint for the dataframe}

\item{...}{A variable list of metadata tables on which this
metadata table depends}
}
\description{
One of the targets in the blueprint drake target chain.
If a metadata file does not exist and the blueprint has
\code{export_metadata} set to \code{TRUE}, then this function will be
added to the drake plan during the blueprint attaching step.
In case the metadata file is accidentally deleted, the \code{.file}
paramter is set to \code{file_out(metadata_path(blueprint))}
so that drake can monitor the metadata file's state.
}
