% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-base.R
\name{eval_checks}
\alias{eval_checks}
\title{Evaluate all checks on a blueprint}
\usage{
eval_checks(..., .env = parent.frame())
}
\arguments{
\item{...}{All quoted check calls}

\item{.env}{The environment in which the calls are evaluated}
}
\description{
Runs all checks -- dataset and variable -- on a blueprint
to determine if a built dataset passes all restrictions.
}
