% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint-provenance.R
\name{vis_table_lineage}
\alias{vis_table_lineage}
\title{Visualize table lineage with visNetwork}
\usage{
vis_table_lineage(...)
}
\arguments{
\item{...}{Arguments passed to \link{load_table_lineage}}
}
\value{
Interactive graph run by visNetwork
}
\description{
Visualize table lineage with visNetwork
}
